requirements.txt
aiohttp==3.9.5
python-dotenv==1.0.1
pandas==2.2.2
numpy==1.26.4
matplotlib==3.8.4
plotly==5.22.0
streamlit==1.36.0
mpi4py==3.1.6
numba==0.59.1
cupy-cuda12x==13.2.0
pyarrow==16.1.0
